# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Staffio HR Management API
  version: 1.0.0
  description: |
    API do sistema de gestão de RH (Staffio). Backend em Go (Gin).

    Autenticação é baseada em sessão (cookie). A maioria dos endpoints exige
    sessão autenticada e os endpoints de escrita respeitam a hierarquia de papéis:
    - `owner`   (dono)  – acesso total
    - `admin`   (admin) – gerencia equipe, edita candidatos
    - `recruit` (recrutador) – edita candidatos e departamentos, sem relatórios/visão geral
    - `viewer`  (visualizador) – somente leitura; bloqueado para métodos não-GET

    Vários endpoints renderizam páginas HTML (servidor). Apenas os endpoints
    prefixados com `/api/` retornam JSON estruturado.
  contact:
    name: Staffio
servers:
  - url: http://localhost:8000
    description: Servidor local
...
paths:
...
  /login:
    get:
      tags: [Pages, Auth]
      summary: Página de login
      description: Renderiza o formulário de login (HTML).
      responses:
        '200':
          description: Página de login renderizada
          content:
            text/html:
              schema: { type: string }
    post:
      tags: [Auth]
      summary: Autenticar usuário
      description: Valida credenciais e cria uma sessão (cookie). Limitado a 10 requisições por minuto por IP.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                  description: Nome de usuário
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '302':
          description: Redireciona para /dashboard em caso de sucesso
        '400':
          description: Campos obrigatórios ausentes
        '401':
          description: Credenciais incorretas
  /register:
    get:
      tags: [Pages, Auth]
      summary: Página de registro
      responses:
        '200':
          description: Página de registro renderizada
          content:
            text/html: { schema: { type: string } }
    post:
      tags: [Auth]
      summary: Registrar nova conta
      description: Cria uma organização e um usuário dono (owner). Limitado a 10 requisições por minuto por IP.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, password, email]
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Conta criada, redireciona para a página de login
        '400':
          description: Campo obrigatório ausente ou usuário já existe
        '500':
          description: Erro interno
...
  /dashboard:
    get:
      tags: [Pages]
      summary: Dashboard
      description: Lista candidatos da organização com busca e paginação (20 por página). Requer autenticação; recrutadores são redirecionados para /employees.
      security:
        - sessionAuth: []
      parameters:
        - name: search
          in: query
          required: false
          description: Filtra por nome ou e-mail
          schema: { type: string }
        - name: all
          in: query
          required: false
          description: "`true` mostra todos (sem limite); `false` limita a 100.000 registros"
          schema: { type: string, enum: ['true', 'false'] }
      responses:
        '200':
          description: Dashboard renderizado
          content:
            text/html: { schema: { type: string } }
        '302':
          description: Redireciona para /login se não autenticado

  # ============================ EMPLOYEES ============================
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Cookie de sessão definido pelo servidor após login
...